Updating 'Kurt' BE branch

+++ database/migrations/2025_02_25_153239_create_test_cases_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
             $table->text('inputData')->nullable();
             $table->text('expectedOutput');
             
-            $table->integer('testCasePoints');
+            $table->float('testCasePoints');
             
             // NEW: Hide from students if true
             $table->boolean('isHidden')->default(false);
